pyproject.toml
[project]
name = "repl-nix-cinemawarehousemanager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "fastapi>=0.115.12",
    "flask-login>=0.6.3",
    "openai>=1.58.1",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "passlib[bcrypt]>=1.7.4",
    "plotly>=5.24.1",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "pytz>=2024.2",
    "streamlit>=1.40.2",
    "uvicorn>=0.34.0",
    "werkzeug>=3.1.3",
    "xlrd>=2.0.1",
]
